Search rework (#61)
Elo   | 250.57 +- 44.99 (95%)
SPRT  | 60.0+0.60s Threads=1 Hash=64MB
LLR   | 1.05 (-2.94, 2.94) [0.50, 2.50]
Games | N: 340 W: 252 L: 42 D: 46
Penta | [3, 4, 33, 40, 90]

Bench: 3746694

+++ senjo/BackgroundCommand.h
@@ -157,7 +157,7 @@ namespace senjo {
         void doWork();
 
     private:
-        bool process(const int depth, const uint64_t expected_leaf_count,
+        bool process(const int16_t depth, const uint64_t expected_leaf_count,
                      uint64_t &leaf_count);
 
         static const std::string _TEST_FILE;
